minor udsource refactor
